small webscrape sample for SFU course outline
we can use the resulting json response as input to construct a url for the actual course outline by using the "values" as the suffix and prefix as "http://www.sfu.ca/outlines.html?"
prefix+suffix will give the course outline. Then we can scrape the outline for information. we can then use that info to query for professor ratings on ratemyprof.